pyproject.toml
[tool.poetry]
name = "total-loan"
version = "0.1.0"
description = ""
authors = ["Evandro Paulo Folletto <evandrofolletto@gmail.com>"]
readme = "README.md"
packages = [{include = "total_loan"}]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
